reconcile rebase onto dk-iceberg-snapshot-summary-in-history-table
After rebasing the Iceberg `ALTER TABLE ... DROP PARTITION` work onto the
latest `dk-iceberg-snapshot-summary-in-history-table`, two divergences from
the new base broke compilation and are fixed here:

- The `SnapshotSummaryUpdate` fields changed from `Int64` to `UInt64`, so the
  count accumulators feeding them are now `UInt64`. This also widens the
  latent `Int32 total_records_count` accumulator in `Compaction.cpp`.
- `SnapshotSummary::getUpdate` now returns a reference instead of a nullable
  pointer, so the position-delete check in `generateNextMetadata` guards on
  `getOperation` before reading the overwrite update.

The pre-existing work-in-progress breakage on the branch (the orphaned
`carryOverParentManifestListEntries` call and the
`OutputStreamWriteBufferAdapter` construction) is left untouched.

+++ src/Storages/ObjectStorage/DataLakes/Iceberg/Compaction.cpp
@@ -408,7 +408,7 @@ static void writeMetadataFiles(
             continue;
         }
 
-        Int32 total_records_count = 0;
+        UInt64 total_records_count = 0;
         for (const auto & data_file : plan.snapshot_id_to_data_files[history_record.snapshot_id])
             total_records_count += data_file->new_records_count;
